feat: 우측 사이드 컴포넌트로 분리
