+++ src/components/PostSideVoteBar/PostSideVoteBar.type.ts
@@ -1,0 +1,5 @@
+export interface PostSideVoteBarProps {
+  className?: string;
+  commentCount: number;
+  votePoints: number;
+}
